+++ tests/test_sampling.py
@@ -7,7 +7,7 @@ import tensorflow as tf
 
 
 import zfit
-from zfit import ztf
+from zfit import ztf, Space
 from zfit.core.sample import accept_reject_sample
 from zfit.util.execution import SessionHolderMixin
 from zfit.core.testing import setup_function, teardown_function, tester
@@ -17,6 +17,18 @@ sigma_true = 1.2
 low, high = -3.8, 2.9
 
 obs1 = 'obs1'
+
+
+class GaussNoAnalyticSampling(zfit.pdf.Gauss):
+
+    def _analytic_sample(self, n, limits: Space):
+        raise NotImplementedError  # HACK do make importance sampling work
+
+
+class UniformNoAnalyticSampling(zfit.pdf.Uniform):
+
+    def _analytic_sample(self, n, limits: Space):
+        raise NotImplementedError  # HACK do make importance sampling work
 
 
 def create_gauss1():
@@ -164,8 +176,9 @@ def test_importance_sampling():
 
     obs_sampler = zfit.Space(obs='obs1', limits=(4.5, 5.5))  # smaller, so pdf is bigger
     obs_pdf = zfit.Space(obs='obs1', limits=(1, 7))
-    gauss_sampler = zfit.pdf.Gauss(mu=mu_sampler, sigma=sigma_sampler, obs=obs_sampler)
-    gauss_pdf = zfit.pdf.Gauss(mu=mu_pdf, sigma=sigma_pdf, obs=obs_pdf)
+
+    gauss_sampler = GaussNoAnalyticSampling(mu=mu_sampler, sigma=sigma_sampler, obs=obs_sampler)
+    gauss_pdf = GaussNoAnalyticSampling(mu=mu_pdf, sigma=sigma_pdf, obs=obs_pdf)
 
     importance_sampling_called = [False]
 
@@ -212,7 +225,7 @@ def test_importance_sampling_uniform():
     low = -3.
     high = 7.
     obs = zfit.Space("obs1", (low, high))
-    uniform = zfit.pdf.Uniform(obs=obs, low=low, high=high)
+    uniform = UniformNoAnalyticSampling(obs=obs, low=low, high=high)
     importance_sampling_called = [False]
 
     class GaussianSampleAndWeights(SessionHolderMixin):
@@ -244,6 +257,7 @@ def test_importance_sampling_uniform():
     # plt.hist(sample_np, bins=40)
     # plt.show()
 
+
 def test_sampling_fixed_eventlimits():
     n_samples1 = 500
     n_samples2 = 400  # just to make sure
@@ -259,7 +273,7 @@ def test_sampling_fixed_eventlimits():
     lower = ((lower,),)
     upper = ((upper,),)
     limits = zfit.core.sample.EventSpace(obs=obs1, limits=(lower, upper))
-    gauss1 = zfit.pdf.Gauss(mu=0.3, sigma=4, obs=zfit.Space(obs=obs1, limits=(-7, 8)))
+    gauss1 = GaussNoAnalyticSampling(mu=0.3, sigma=4, obs=zfit.Space(obs=obs1, limits=(-7, 8)))
 
     sample = gauss1.sample(n=n_samples_tot, limits=limits)
     sample_np = zfit.run(sample)
@@ -272,4 +286,4 @@ def test_sampling_fixed_eventlimits():
     assert all(sample_np[n_samples2:n_samples3] <= upper3)
     with pytest.raises(ValueError,
                        match="are incompatible"):  # cannot use the exact message, () are regex syntax... bug in pytest
-        _ = gauss1.sample(n=n_samples_tot + 1, limits=limits)
+        _ = gauss1.sample(n=n_samples_tot + 1, limits=limits)  # TODO(Mayou36): catch analytic integral